feat: create common package
