+++ Cargo.toml
@@ -19,6 +19,7 @@ incremental = false
 overflow-checks = true
 
 [workspace.dependencies]
+common = { path = "./packages/common" }
 cosmwasm-schema = "1.2"
 cosmwasm-std = { version = "1.2", features = ["staking"] }
 cosmwasm-storage = "1.1.3"
