+++ packages/common/src/types.rs
@@ -1,0 +1,4 @@
+pub type Bytes = Vec<u8>;
+pub type Commitment = Hash;
+pub type Memo = Vec<u8>;
+pub type Hash = String;
