+++ packages/integration-tests/Cargo.toml
@@ -27,6 +27,7 @@ optimize = """docker run --rm -v "$(pwd)":/code \
 
 # TODO: check if all of these are needed
 [dependencies]
+common = { workspace = true }
 cosmwasm-schema = { workspace = true }
 cosmwasm-std = { workspace = true }
 cw-storage-plus = { workspace = true }
