+++ packages/proxy/Cargo.toml
@@ -26,10 +26,10 @@ optimize = """docker run --rm -v "$(pwd)":/code \
 """
 
 [dependencies]
+common = { workspace = true }
 cosmwasm-schema = { workspace = true }
 cosmwasm-std = { workspace = true }
 cw-storage-plus = { workspace = true }
 cw2 = { workspace = true }
-seda-chain-contracts = { path = "../seda-chain-contracts" }
 serde = { workspace = true }
 thiserror = { workspace = true }
