+++ packages/seda-chain-contracts/src/bin/schema.rs
@@ -1,6 +1,7 @@
 use cosmwasm_schema::write_api;
 
-use seda_chain_contracts::msg::{ExecuteMsg, InstantiateMsg, QueryMsg};
+use common::msg::{ExecuteMsg, QueryMsg};
+use seda_chain_contracts::msg::InstantiateMsg;
 
 fn main() {
     write_api! {
