+++ packages/seda-chain-contracts/src/types.rs
@@ -1,6 +1,2 @@
-pub type Hash = String;
 pub type Input = Vec<u8>;
-pub type Memo = Vec<u8>;
 pub type PayloadItem = Vec<u8>;
-pub type Commitment = Hash;
-pub type Bytes = Vec<u8>;
